# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Etrade API",
    "description": "API documentation for Etrade application",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://megadeal.dev",
      "description": "Local server"
    }
  ],
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful registration",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successful"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AuthResponse"
                    }
                  }
                },
                "example": {
                  "code": 200,
                  "message": "Successful",
                  "data": {
                    "token": "gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c",
                    "expiresIn": 3600,
                    "user": {
                      "id": 1,
                      "username": "john_doe",
                      "email": "john.doe@example.com",
                      "fullName": "John Doe",
                      "phone": "+84 909 123 456",
                      "address": "123 Nguyen Trai Street, District 1, Ho Chi Minh City",
                      "role": "USER",
                      "createdAt": "2026-05-15T10:30:00Z",
                      "updatedAt": "2026-05-15T14:45:00Z"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict - User already exists or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Username already exists"
                    },
                    "code": {
                      "type": "integer",
                      "example": 409
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successful"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AuthResponse"
                    }
                  }
                },
                "example": {
                  "code": 200,
                  "message": "Successful",
                  "data": {
                    "token": "gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c",
                    "expiresIn": 3600,
                    "user": {
                      "id": 1,
                      "username": "john_doe",
                      "email": "john.doe@example.com",
                      "fullName": "John Doe",
                      "phone": "+84 909 123 456",
                      "address": "123 Nguyen Trai Street, District 1, Ho Chi Minh City",
                      "role": "USER",
                      "createdAt": "2026-05-15T10:30:00Z",
                      "updatedAt": "2026-05-15T14:45:00Z"
                    }
                  }
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "description": "Access token cookie",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "integer",
                      "example": 409
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get current user information",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserEntity"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/google": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login with Google",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Google login",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successful"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AuthGoogleResponse"
                    }
                  }
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "description": "Access token cookie",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Google login failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "integer",
                      "example": 409
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/product/{id}": {
      "get": {
        "tags": ["Products"],
        "summary": "Get product by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Product ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Product details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{categoryName}": {
      "get": {
        "tags": ["Products"],
        "summary": "Get products by category name",
        "parameters": [
          {
            "name": "categoryName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Category name"
          }
        ],
        "responses": {
          "200": {
            "description": "List of products in category",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Username for registration"
          },
          "password": {
            "type": "string",
            "description": "Password for registration"
          },
          "email": {
            "type": "string",
            "description": "Email address"
          },
          "fullName": {
            "type": "string",
            "description": "Full name"
          },
          "phone": {
            "type": "string",
            "description": "Phone number"
          },
          "address": {
            "type": "string",
            "description": "Address"
          }
        },
        "required": ["username", "password", "email"]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Username"
          },
          "password": {
            "type": "string",
            "description": "Password"
          }
        },
        "required": ["username", "password"]
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "JWT access token",
            "example": "JhbGciOiJIUzIRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
          },
          "expiresIn": {
            "type": "integer",
            "description": "Token expiration time in seconds",
            "example": 3600
          },
          "user": {
            "$ref": "#/components/schemas/UserEntity"
          }
        }
      },
      "AuthGoogleResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT access token from Google login",
            "example": "KxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
          },
          "expiresIn": {
            "type": "integer",
            "description": "Token expiration time in seconds",
            "example": 3600
          },
          "user": {
            "$ref": "#/components/schemas/UserEntity"
          }
        }
      },
      "UserEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "User ID",
            "example": 1
          },
          "username": {
            "type": "string",
            "description": "Username",
            "example": "john_doe"
          },
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "john.doe@example.com"
          },
          "fullName": {
            "type": "string",
            "description": "Full name",
            "example": "John Doe"
          },
          "phone": {
            "type": "string",
            "description": "Phone number",
            "example": "+84 909 123 456"
          },
          "address": {
            "type": "string",
            "description": "Address",
            "example": "123 Nguyen Trai Street, District 1, Ho Chi Minh City"
          },
          "role": {
            "type": "string",
            "description": "User role (USER, ADMIN)",
            "example": "USER"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp",
            "example": "2026-05-15T10:30:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp",
            "example": "2026-05-15T14:45:00Z"
          }
        }
      },
      "ProductResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Product ID",
            "example": 101
          },
          "name": {
            "type": "string",
            "description": "Product name",
            "example": "iPhone 15 Pro Max"
          },
          "description": {
            "type": "string",
            "description": "Product description",
            "example": "Latest Apple smartphone with titanium design and advanced camera system"
          },
          "price": {
            "type": "number",
            "format": "float",
            "description": "Product price",
            "example": 1099.99
          },
          "stock": {
            "type": "integer",
            "description": "Available stock quantity",
            "example": 50
          },
          "category": {
            "type": "string",
            "description": "Product category",
            "example": "Smartphones"
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductImageResponse"
            }
          },
          "colors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ColorResponse"
            }
          },
          "sizes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SizeResponse"
            }
          }
        }
      },
      "ProductImageResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "url": {
            "type": "string",
            "description": "Image URL",
            "example": "https://example.com/images/iphone15-pro-main.jpg"
          }
        }
      },
      "ColorResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Color name",
            "example": "Black"
          },
          "code": {
            "type": "string",
            "description": "Color code (hex)",
            "example": "#000000"
          }
        }
      },
      "SizeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Size name",
            "example": "Large"
          }
        }
      },
...
    }
  }
}